added support for multiple regex patterns
This allows for a user to call user_input as a list with multiple regex patterns which are then called sequentially. user_input can still be called as a string, so no previous functionality is lost. I also set up the [Patterns not Found:] box to show the patterns which were not found since the user may not know which regex pattern failed if they run multiple.