Polishing the named_patterns call
added support for calling named patterns as well as raw regexp when creating a user_input list. This also supports lists of regexp within the named_patterns and will rebuild the user_insert list such that user order is maintained across both the initial user_insert list and any regexp that are in the named_patterns list.