exclude unittests from sublime-packages
Unittests and their configuration files are not needed by ST, but only for development purposes. Therefore they don't need to be part of the sublime-package. This commit therefore proposes the use of an .gitattribute file to exclude those files from being added to the zip archive, which is downloaded by Package Control upon install/update.